+++ R/helpfunctions.R
@@ -3,6 +3,10 @@
 magrittr::`%>%`
 
 
+# do we have a stan-model?
+is.stan <- function(x) inherits(x, c("stanreg", "stanfit", "brmsfit"))
+
+
 #' @importFrom dplyr quos select
 get_dot_data <- function(x, qs) {
   if (isempty(qs))
@@ -166,15 +170,10 @@ isempty <- function(x, first.only = TRUE) {
 }
 
 
+
+#' @importFrom snakecase to_any_case
 convert_case <- function(lab, case) {
   if (!is.null(case) && !is.null(lab)) {
-
-    # check if package available
-    if (!requireNamespace("snakecase", quietly = TRUE)) {
-      message("Package `snakecase` required for case-conversion.")
-      return(lab)
-    }
-
     snakecase::to_any_case(
       lab,
       case = case,