+++ NEWS.md
@@ -1,3 +1,14 @@
+# sjlabelled 1.0.4
+
+## Changes to functions
+
+* `get_term_labels()` now includes variable names for factors with numeric factor levels only (and not only return the numeric level as term label).
+
+## Bug fixes
+
+* Fixed bug for `as_label()`, when `x` was a character vector and argument `drop.levels` was `TRUE`.
+* Fixed issue for *lme* and *gls* objects in `get_term_labels()` and `get_dv_labels()`.
+
 # sjlabelled 1.0.3
 
 ## General